api/requirements.txt
aniso8601==3.0.2
bcrypt==3.1.4
cffi==1.11.5
Click==7.0
coverage==4.5.1
Flask==1.0.2
Flask-Cors==3.0.6
Flask-JWT-Extended==3.13.1
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysql-connector==2.1.6
passlib==1.7.1
pycparser==2.19
PyJWT==1.6.4
pytz==2018.5
PyYAML==3.13
six==1.11.0
SQLAlchemy==1.2.12
Werkzeug==0.14.1
